final project/requirements.txt
asgiref==3.3.1
attrs==19.3.0
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.7
django-environ==0.4.5
django-heroku==0.3.1
django-markdown2==0.3.1
django-multiselectfield==0.1.12
gunicorn==20.1.0
idna==2.10
jellyfish==0.8.2
lib50==1.1.11
markdown2==2.4.0
pexpect==4.8.0
Pillow==8.2.0
psycopg2==2.8.6
ptyprocess==0.7.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
sqlparse==0.4.1
termcolor==1.1.0
urllib3==1.26.3
whitenoise==5.2.0
